+++ celescope/__init__.py
@@ -1,7 +1,7 @@
 import os
 from collections import OrderedDict
 
-__VERSION__ = "2.0.5"
+__VERSION__ = "2.0.6"
 __version__ = __VERSION__
 
 ASSAY_LIST = [